requirements.txt
astroid==2.0.4
certifi==2018.4.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.0.7
django-cors-headers==2.4.0
django-oauth-toolkit==1.2.0
djangorestframework==3.8.2
idna==2.7
isort==4.3.4
itypes==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.1.1
mccabe==0.6.1
oauthlib==2.1.0
Pillow==5.2.0
psycopg2==2.7.5
psycopg2-binary==2.7.5
pylint==2.1.1
python-magic==0.4.15
pytz==2018.5
raven==6.9.0
requests==2.19.1
six==1.11.0
typed-ast==1.1.0
uritemplate==3.0.0
urllib3==1.23
wrapt==1.10.11
